requirements.txt
flask
faiss-cpu
sentence-transformers
python-dotenv
